% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseglm-package.R
\docType{package}
\name{sparseglm-package}
\alias{sparseglm}
\alias{sparseglm-package}
\title{sparseglm: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zamorarr/sparseglm}
  \item Report bugs at \url{https://github.com/zamorarr/sparseglm/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
